veille-backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
websockets>=12.0
pydantic-settings>=2.0
python-dotenv>=1.0
supabase>=2.0
twilio>=9.0
openai>=1.0
httpx>=0.27
gradium
audioop-lts
